Seperate request and access token + refactor code
Seperated request with request with request token and actual access
token.
Refactoring Code for parsing request bodys.
